% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{schema_to_jobj}
\alias{schema_to_jobj}
\title{Convert an Avro schema to a Spark StructType Java object}
\usage{
schema_to_jobj(sc, schema)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{schema}{Character or list. An Avro schema (JSON string or parsed list).}
}
\value{
A Spark \code{StructType} Java object (\code{spark_jobj}).
}
\description{
Parses an Avro JSON schema (character or list) and builds the corresponding
Spark \code{StructType} via \code{sparklyr.nested} type constructors. This is
useful for supplying a \code{schema} argument to
\code{\link[sparklyr]{spark_read_parquet}} or
\code{\link[sparklyr]{spark_read_avro}}.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
schema <- '{"type":"record","name":"X","fields":[{"name":"a","type":"string"}]}'
jobj <- schema_to_jobj(sc, schema)
df <- sparklyr::spark_read_parquet(sc, path = "data.parquet", schema = jobj)
}

}
