% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sdf_schema_json}
\alias{sdf_schema_json}
\alias{sdf_schema_viewer}
\title{Work with the schema}
\usage{
sdf_schema_json(x, simplify = FALSE, append_complex_type = TRUE)

sdf_schema_viewer(x, simplify = TRUE, append_complex_type = TRUE)
}
\arguments{
\item{x}{An \code{R} object wrapping, or containing, a Spark DataFrame.}

\item{simplify}{Logical. If \code{TRUE} then the schema will be folded into itself such that
\code{{"name" : "field1", "type" : {"type" : "array", "elementType" : "string", "containsNull" : true},
 "nullable" : true, "metadata" : { } }} will be rendered simply \code{{"field1 (array)" : "[string]"}}}

\item{append_complex_type}{Logical. This only matters if \code{simplify=TRUE}. In that case indicators will
be included in the return value for array and struct types.}
}
\description{
Work with the schema
}
\seealso{
\code{\link[sparklyr]{sdf_schema}}
}
