% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{struct_type}
\alias{struct_type}
\alias{struct_field}
\alias{array_type}
\alias{binary_type}
\alias{boolean_type}
\alias{byte_type}
\alias{date_type}
\alias{double_type}
\alias{float_type}
\alias{integer_type}
\alias{numeric_type}
\alias{map_type}
\alias{string_type}
\alias{character_type}
\alias{timestamp_type}
\title{Spark Data Types}
\usage{
struct_type(sc, struct_fields)

struct_field(sc, name, data_type, nullable = FALSE)

array_type(sc, data_type, nullable = FALSE)

binary_type(sc)

boolean_type(sc)

byte_type(sc)

date_type(sc)

double_type(sc)

float_type(sc)

integer_type(sc)

numeric_type(sc)

map_type(sc, key_type, value_type, nullable = FALSE)

string_type(sc)

character_type(sc)

timestamp_type(sc)
}
\description{
These function support supplying a spark read schema. This is particularly useful
when reading data with nested arrays when you are not interested in several of 
the nested fields.
}
